% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_demo.R
\docType{data}
\name{site_demo}
\alias{site_demo}
\title{Data of site information in the demo data set}
\format{
A tibble with 1000 rows and 13 variables:
\describe{
\item{SITE_ID}{chr contains the unique identification for each site}
\item{LAT}{num contains the site geographic coordinate latitude measured in degree}
\item{LONG}{num contains the site geographic coordinate longitude measured in degree}
\item{PM}{chr contains the site parent material information }
\item{CSSC_CODE}{chr contains the site abbrevated CSSC subgroup code e.g., E.DYB}
\item{ORDER}{chr contains the site CSSC soil order name}
\item{GG}{chr contains the site CSSC soil great group name}
\item{GGSG}{chr contains the site CSSC soil subgroup name}
\item{PROJ_ID}{chr contains the site original project ID}
\item{PROVINCE}{chr contains the site located province in Canada}
\item{YEAR}{int contains the site observation data (Year)}
\item{DATA_SOURCE}{chr contains the site data source}
\item{DATA_MODIFIED}{chr contains the data modification date}
}
}
\source{
\url{local_server_SOIL_2020}
}
\usage{
site_demo
}
\description{
Site information are collected by site, and unique site ID is given for individual site.
}
\keyword{datasets}
